apps/ai-curator/requirements.txt
# AI Curator - Python Dependencies
# FastAPI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Database connections
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
clickhouse-connect>=0.7.0
redis>=5.0.1

# ML & Data Science
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
scikit-surprise>=1.1.3

# Text processing & embeddings
sentence-transformers>=2.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
